Add rating diff; move to i16 for data
Move to i16 is to save memory when processing a large number of games

+++ src/database.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 #[derive(Default, Debug, Clone)]
 pub struct Database {
     pub children: HashMap<String, Database>,
-    pub data: Vec<i32>,
+    pub data: Vec<i16>,
 }
 
 impl Database {
